requirements.txt
pymongo
psycopg2-binary
pdfplumber
python-docx
transformers
torch
tqdm
python-dotenv
flask
flask-cors
groq
supabase